Import: move Suunto .json/.fit pairing out of mainwindow.cpp
mikeller pointed out in review that the .json/.fit pairing logic added
for the Suunto Ocean gas-mix fix lived in mainwindow.cpp, even though
it is entirely Suunto-specific. Move the SuuntoPair struct and the
pairing/reading loop into core/import-suunto-json.cpp as a new
suunto_json_fit_pair_import() function, declared in core/file.h next
to suunto_json_import(). mainwindow.cpp's importFiles() now just calls
that function and no longer needs to know about Suunto file formats at
all.

No behaviour change: the moved code is unchanged except for reading
readfile()/report_info() from its new home and using QFile::encodeName()
directly instead of mainwindow.cpp's local encodeFileName() helper.
Verified with the existing and new Suunto test cases in testparse.cpp,
all of which still pass.
