+++ core/file.h
@@ -46,6 +46,13 @@ extern int logtrak_import(const std::string &mem, struct divelog *log);
 extern int scubapro_asd_import(const std::string &mem, struct divelog *log);
 extern int fit_file_import(const std::string &buffer, struct divelog *log);
 extern int suunto_json_import(const std::string &buffer, const std::string &fit_buffer, struct divelog *log);
+// AI-generated (Claude)
+// Detects .json/.fit pairs sharing a base name among fileNames, as exported
+// together by the Suunto app for one dive, and imports each complete pair as
+// a single dive via suunto_json_import(). Marks consumed[i] for every file
+// that was part of a complete pair; unpaired files are left for the caller's
+// normal per-file import path.
+extern void suunto_json_fit_pair_import(const std::vector<std::string> &fileNames, std::vector<bool> &consumed, struct divelog *log);
 
 extern int try_to_open_cochran(const char *filename, std::string &mem, struct divelog *log);
 extern int try_to_open_liquivision(const char *filename, std::string &mem, struct divelog *log);
