+++ core/import-suunto-json.cpp
@@ -26,6 +26,10 @@
 #include <cstdint>
 #include <cstdio>
 
+#include <map>
+
+#include <QFile>
+#include <QFileInfo>
 #include <QJsonArray>
 #include <QJsonDocument>
 #include <QJsonObject>
@@ -37,6 +41,7 @@
 #include "divesite.h"
 #include "errorhelper.h"
 #include "extradata.h"
+#include "file.h"
 #include "gettext.h"
 #include "libdivecomputer.h"
 #include "sample.h"
@@ -710,4 +715,49 @@ int suunto_json_import(const std::string &buffer, const std::string &fit_buffer,
 
 	log->dives.record_dive(std::move(d));
 	return 1;
+}
+
+// AI-generated (Claude)
+void suunto_json_fit_pair_import(const std::vector<std::string> &fileNames, std::vector<bool> &consumed, struct divelog *log)
+{
+	/* The Suunto app exports a .json and a .fit file with the same base
+	 * name for the same dive. If the caller selected both, merge them into
+	 * a single suunto_json_import() call so the JSON's richer profile and
+	 * the FIT's gas mix/gradient factors end up on one dive, instead of
+	 * two unrelated dives. Only exact, case-insensitive base-name matches
+	 * within fileNames are paired -- no scanning the filesystem for a
+	 * sibling file that wasn't explicitly selected. */
+	struct SuuntoPair {
+		size_t jsonIdx = std::string::npos;
+		size_t fitIdx = std::string::npos;
+	};
+	std::map<QString, SuuntoPair> pairs;
+	for (size_t i = 0; i < fileNames.size(); ++i) {
+		QFileInfo fi(QString::fromStdString(fileNames[i]));
+		QString suffix = fi.suffix().toLower();
+		if (suffix != "json" && suffix != "fit")
+			continue;
+		SuuntoPair &pair = pairs[fi.completeBaseName().toLower()];
+		if (suffix == "json")
+			pair.jsonIdx = i;
+		else
+			pair.fitIdx = i;
+	}
+
+	for (const auto &[stem, pair]: pairs) {
+		if (pair.jsonIdx == std::string::npos || pair.fitIdx == std::string::npos)
+			continue; // not a complete pair -- both files import individually in the caller
+		std::string jsonEncoded = QFile::encodeName(QString::fromStdString(fileNames[pair.jsonIdx])).toStdString();
+		std::string fitEncoded = QFile::encodeName(QString::fromStdString(fileNames[pair.fitIdx])).toStdString();
+		auto [jsonBuf, jerr] = readfile(jsonEncoded.c_str());
+		auto [fitBuf, ferr] = readfile(fitEncoded.c_str());
+		if (jerr <= 0)
+			continue; // couldn't read the JSON -- fall through to the normal per-file import in the caller
+		if (ferr <= 0)
+			report_info("Suunto import: could not read paired FIT file '%s', importing gas mix from JSON as-is",
+				    fileNames[pair.fitIdx].c_str());
+		suunto_json_import(jsonBuf, ferr > 0 ? fitBuf : std::string(), log);
+		consumed[pair.jsonIdx] = true;
+		consumed[pair.fitIdx] = true;
+	}
 }
